requirements.txt
langchain
openai
langchain-openai
boto3
python-dotenv
langchain-core
langgraph
